Mv gitlab.ts fully into stable code, update fn signatures
